Mod: port off flecs onto native MafiaNet replication (#17)
The framework replaced flecs + the streamer/message entity-sync layer
with native ReplicaManager3 + RPC4 (#201/#205). Port HogwartsMP onto it:

- Humans are NetworkEntity subclasses: a shared HumanEntity (spawnProfile
  + nickname in the construction snapshot) and a client ClientHuman that
  binds the local player when owned or spawns a student proxy otherwise.
  Both register the same "HogwartsMP::Human" type with the EntityRegistry.
- The server builds avatars in onPlayerConnect via ReplicationManager
  (CreateEntity/SetOwner/SetViewer); World.spawnHuman spawns server-owned
  NPCs. Per-entity spawn/update/despawn messages are gone.
- Chat moves to the framework's built-in ChatMessage RPC and chat/connect
  callbacks; weather is a plain WeatherState + migrated SetWeather RPC.
- Scripting Human now derives the framework's NetworkID-based Player
  builtin; the old flecs Entity builtin is removed.
- Size the interest grid for Hogwarts' cm-scale coordinates.
- Rename the SDK's SHA1_WORKSPACE_BLOCK to avoid a clash with mafianet's
  DR_SHA1.h, now pulled in transitively by the networking headers.

+++ code/client/src/core/hooks/player_controller.cpp
@@ -31,7 +31,7 @@ void APlayerController_BeginPlay_Hook(void *pThis) {
 typedef void(__fastcall *APlayerController_EndPlay_t)(void *, EndPlayReason);
 APlayerController_EndPlay_t APlayerController_EndPlay_original = nullptr;
 void APlayerController_EndPlay_Hook(void *pThis, EndPlayReason reason) {
-    Framework::Logging::GetLogger("Hooks")->info("APlayerController::EndPlay with reason {}", reason);
+    Framework::Logging::GetLogger("Hooks")->info("APlayerController::EndPlay with reason {}", static_cast<int>(reason));
     APlayerController_EndPlay_original(pThis, reason);
 }
